issue #2 変更: 天気情報表示部分とボタン部分のウィジェットを別ファイルのコンポーネントとして書き出す
