+++ lib/presentation/weather/components/action_buttons.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class ActionButtons extends StatelessWidget {
+  const ActionButtons({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+      children: [
+        Flexible(
+          child: TextButton(
+            onPressed: () {
+              // TODO: ボタン押下時に何かする
+            },
+            child: const Text('Close'),
+          ),
+        ),
+        Flexible(
+          child: TextButton(
+            onPressed: () {
+              // TODO: ボタン押下時に何かする
+            },
+            child: const Text('Reload'),
+          ),
+        ),
+      ],
+    );
+  }
+}
